+++ src/peer_manager/mod.rs
@@ -889,6 +889,8 @@ impl PeerManager {
         transport.max_prefixes_ipv6 = config.max_prefixes_ipv6;
         transport.max_prefixes_received_ipv4 = config.max_prefixes_received_ipv4;
         transport.max_prefixes_received_ipv6 = config.max_prefixes_received_ipv6;
+        transport.max_prefix_action = config.max_prefix_action;
+        transport.max_prefix_warning_percent = config.max_prefix_warning_percent;
         transport.peer_group.clone_from(&config.peer_group);
         transport.md5_password.clone_from(&config.md5_password);
         transport.tcp_ao.clone_from(&config.tcp_ao);